♻️ refactor(webapp): split shared form and alert primitives
