+++ apps/webapp/src/components/ui/alert-title.tsx
@@ -1,0 +1,15 @@
+import type * as React from "react";
+
+import { cn } from "@/lib/utils";
+
+function AlertTitle({ className, ...props }: React.ComponentProps<"div">) {
+	return (
+		<div
+			data-slot="alert-title"
+			className={cn("col-start-2 line-clamp-1 min-h-4 font-medium tracking-tight", className)}
+			{...props}
+		/>
+	);
+}
+
+export { AlertTitle };
