+++ apps/webapp/src/components/ui/alert.tsx
@@ -1,6 +1,8 @@
 import { cva, type VariantProps } from "class-variance-authority";
 import type * as React from "react";
 
+import { AlertDescription } from "@/components/ui/alert-description";
+import { AlertTitle } from "@/components/ui/alert-title";
 import { cn } from "@/lib/utils";
 
 const alertVariants = cva(
@@ -29,29 +31,6 @@ function Alert({
 			data-slot="alert"
 			role="alert"
 			className={cn(alertVariants({ variant }), className)}
-			{...props}
-		/>
-	);
-}
-
-function AlertTitle({ className, ...props }: React.ComponentProps<"div">) {
-	return (
-		<div
-			data-slot="alert-title"
-			className={cn("col-start-2 line-clamp-1 min-h-4 font-medium tracking-tight", className)}
-			{...props}
-		/>
-	);
-}
-
-function AlertDescription({ className, ...props }: React.ComponentProps<"div">) {
-	return (
-		<div
-			data-slot="alert-description"
-			className={cn(
-				"text-muted-foreground col-start-2 grid justify-items-start gap-1 text-sm [&_p]:leading-relaxed",
-				className,
-			)}
 			{...props}
 		/>
 	);
