+++ apps/webapp/src/components/ui/breadcrumb.tsx
@@ -53,10 +53,8 @@ function BreadcrumbPage({ className, ...props }: React.ComponentProps<"span">) {
 	return (
 		<span
 			aria-current="page"
-			aria-disabled="true"
 			className={cn("font-normal text-foreground", className)}
 			data-slot="breadcrumb-page"
-			role="link"
 			{...props}
 		/>
 	);
