+++ apps/webapp/src/components/ui/button-group.tsx
@@ -8,13 +8,12 @@ function ButtonGroup({
 	className,
 	orientation,
 	...props
-}: React.ComponentProps<"div"> & VariantProps<typeof buttonGroupVariants>) {
+}: React.ComponentProps<"fieldset"> & VariantProps<typeof buttonGroupVariants>) {
 	return (
-		<div
-			role="group"
+		<fieldset
 			data-slot="button-group"
 			data-orientation={orientation}
-			className={cn(buttonGroupVariants({ orientation }), className)}
+			className={cn("m-0 min-w-0 border-0 p-0", buttonGroupVariants({ orientation }), className)}
 			{...props}
 		/>
 	);
