+++ apps/webapp/src/components/ui/field.tsx
@@ -77,13 +77,12 @@ function Field({
 	className,
 	orientation = "vertical",
 	...props
-}: React.ComponentProps<"div"> & VariantProps<typeof fieldVariants>) {
+}: React.ComponentProps<"fieldset"> & VariantProps<typeof fieldVariants>) {
 	return (
-		<div
-			role="group"
+		<fieldset
 			data-slot="field"
 			data-orientation={orientation}
-			className={cn(fieldVariants({ orientation }), className)}
+			className={cn("m-0 min-w-0 border-0 p-0", fieldVariants({ orientation }), className)}
 			{...props}
 		/>
 	);
