+++ apps/webapp/src/components/ui/input-group.tsx
@@ -1,18 +1,18 @@
 "use client";
 
-import { cva, type VariantProps } from "class-variance-authority";
 import type * as React from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Textarea } from "@/components/ui/textarea";
+import { InputGroupAddon } from "@/components/ui/input-group-addon";
+import { InputGroupButton } from "@/components/ui/input-group-button";
+import { InputGroupInput, InputGroupTextarea } from "@/components/ui/input-group-control";
+import { InputGroupText } from "@/components/ui/input-group-text";
 import { cn } from "@/lib/utils";
 
-function InputGroup({ className, ...props }: React.ComponentProps<"div">) {
+function InputGroup({ className, ...props }: React.ComponentProps<"fieldset">) {
 	return (
-		<div
+		<fieldset
 			data-slot="input-group"
-			role="group"
 			className={cn(
+				"m-0 min-w-0 p-0",
 				"group/input-group border-input bg-card dark:bg-input/30 relative flex w-full items-center rounded-md border shadow-xs transition-[color,box-shadow] outline-none",
 				"h-9 min-w-0 has-[>textarea]:h-auto",
 
@@ -28,131 +28,6 @@ function InputGroup({ className, ...props }: React.ComponentProps<"div">) {
 				// Error state.
 				"has-[[data-slot][aria-invalid=true]]:ring-destructive/20 has-[[data-slot][aria-invalid=true]]:border-destructive dark:has-[[data-slot][aria-invalid=true]]:ring-destructive/40",
 
-				className,
-			)}
-			{...props}
-		/>
-	);
-}
-
-const inputGroupAddonVariants = cva(
-	"text-muted-foreground flex h-auto cursor-text items-center justify-center gap-2 py-1.5 text-sm font-medium select-none [&>svg:not([class*='size-'])]:size-4 [&>kbd]:rounded-[calc(var(--radius)-5px)] group-data-[disabled=true]/input-group:opacity-50",
-	{
-		variants: {
-			align: {
-				"inline-start": "order-first pl-3 has-[>button]:ml-[-0.45rem] has-[>kbd]:ml-[-0.35rem]",
-				"inline-end": "order-last pr-3 has-[>button]:mr-[-0.45rem] has-[>kbd]:mr-[-0.35rem]",
-				"block-start":
-					"order-first w-full justify-start px-3 pt-3 [.border-b]:pb-3 group-has-[>input]/input-group:pt-2.5",
-				"block-end":
-					"order-last w-full justify-start px-3 pb-3 [.border-t]:pt-3 group-has-[>input]/input-group:pb-2.5",
-			},
-		},
-		defaultVariants: {
-			align: "inline-start",
-		},
-	},
-);
-
-function InputGroupAddon({
-	className,
-	align = "inline-start",
-	...props
-}: React.ComponentProps<"div"> & VariantProps<typeof inputGroupAddonVariants>) {
-	function focusInput(element: HTMLDivElement) {
-		element.parentElement?.querySelector("input")?.focus();
-	}
-
-	return (
-		<div
-			role="button"
-			tabIndex={0}
-			aria-label="Focus input"
-			data-slot="input-group-addon"
-			data-align={align}
-			className={cn(inputGroupAddonVariants({ align }), className)}
-			onClick={(e) => {
-				if ((e.target as HTMLElement).closest("button")) {
-					return;
-				}
-				focusInput(e.currentTarget);
-			}}
-			onKeyDown={(e) => {
-				if (e.key !== "Enter" && e.key !== " ") {
-					return;
-				}
-				e.preventDefault();
-				focusInput(e.currentTarget);
-			}}
-			{...props}
-		/>
-	);
-}
-
-const inputGroupButtonVariants = cva("text-sm shadow-none flex gap-2 items-center", {
-	variants: {
-		size: {
-			xs: "h-6 gap-1 px-2 rounded-[calc(var(--radius)-5px)] [&>svg:not([class*='size-'])]:size-3.5 has-[>svg]:px-2",
-			sm: "h-8 px-2.5 gap-1.5 rounded-md has-[>svg]:px-2.5",
-			"icon-xs": "size-6 rounded-[calc(var(--radius)-5px)] p-0 has-[>svg]:p-0",
-			"icon-sm": "size-8 p-0 has-[>svg]:p-0",
-		},
-	},
-	defaultVariants: {
-		size: "xs",
-	},
-});
-
-function InputGroupButton({
-	className,
-	type = "button",
-	variant = "ghost",
-	size = "xs",
-	...props
-}: Omit<React.ComponentProps<typeof Button>, "size"> &
-	VariantProps<typeof inputGroupButtonVariants>) {
-	return (
-		<Button
-			type={type}
-			data-size={size}
-			variant={variant}
-			className={cn(inputGroupButtonVariants({ size }), className)}
-			{...props}
-		/>
-	);
-}
-
-function InputGroupText({ className, ...props }: React.ComponentProps<"span">) {
-	return (
-		<span
-			className={cn(
-				"text-muted-foreground flex items-center gap-2 text-sm [&_svg]:pointer-events-none [&_svg:not([class*='size-'])]:size-4",
-				className,
-			)}
-			{...props}
-		/>
-	);
-}
-
-function InputGroupInput({ className, ...props }: React.ComponentProps<"input">) {
-	return (
-		<Input
-			data-slot="input-group-control"
-			className={cn(
-				"flex-1 rounded-none border-0 bg-transparent shadow-none focus-visible:ring-0 dark:bg-transparent",
-				className,
-			)}
-			{...props}
-		/>
-	);
-}
-
-function InputGroupTextarea({ className, ...props }: React.ComponentProps<"textarea">) {
-	return (
-		<Textarea
-			data-slot="input-group-control"
-			className={cn(
-				"flex-1 resize-none rounded-none border-0 bg-transparent py-3 shadow-none focus-visible:ring-0 dark:bg-transparent",
 				className,
 			)}
 			{...props}
